requirements.txt
algorithmia
flask >= 1.1
pillow
python-resize-image
pyjwt
pymongo
